refactor: extract compact-mode formatting into _CompactFormatMixin
Second decomposition cut. Moves the compact-response size/shape helpers —
_resolve_compact_budget / _wrap_retrieved_content / _truncate_search_snippets
/ _truncation_footer / _cap_response_size / _strip_markdown_links plus their
budget-profile, content-fence, markdown-regex and atomic-intent constants —
into openzim_mcp/compact_format.py as _CompactFormatMixin.

These six methods call no other handler methods, so the mixin is fully
self-contained (no TYPE_CHECKING cross-declarations). Constants stay class
attributes so existing `self._CONTENT_FENCE_OVERHEAD` / `self._cap_response_size`
call sites resolve unchanged via the MRO. _strip_markdown_links was a
staticmethod that hard-referenced `SimpleToolsHandler._MARKDOWN_*` by class
name (which can't work from a mixin module); converted to a classmethod using
`cls._MARKDOWN_*` — behavior identical since every caller invokes it as
`self._strip_markdown_links(...)`.

Pure code movement, no behavior change. simple_tools.py 5,623 → 5,358
(5,896 → 5,358 across the two mixin cuts). Full suite green (2623 passed);
mypy/flake8/isort clean.
